+++ lib/Player-Core.ts
@@ -1,7 +1,7 @@
 
 import { PlayerOptions } from "./Player-Options";
 import { throttle } from "./Player-Tools";
-import { PlayerEventEmitter, type ValidEvents, type Listener } from "./Player-Event";
+import { PlayerEventEmitter, type ValidEvents, type Listener, PlayerEvent } from "./Player-Event";
 import { Representation } from "./Player-Representation";
 import { PlayerError } from "./Player-Error";
 import type { Processor, ProcessorType, RepType } from "./Player-Processor";
@@ -19,9 +19,10 @@ const NullProcessor = new class implements Processor {
 
 const processorList = new Map<ProcessorType["name"], ProcessorType["asyncFunctionProcessorInstance"]>()
 
-const getRepList = (repType: RepType, processor?: Processor) =>
-      processor?.get(repType).map(rep => ({ ...rep, switch(options: SwitchRepOptions = { switchMode: "soft" }) { 
-        processor.switch(repType, rep, options)    
+const getRepList = (repType: RepType, processor: Processor,currentTime:number) =>
+      processor.get(repType).map(rep => ({ ...rep, switch(options: SwitchRepOptions = { switchMode: "soft" }) { 
+      //  console.log("mpd-getRepList",options);
+        processor.switch(repType, rep,currentTime, options)    
     } })) || []
 
 /**
@@ -35,34 +36,41 @@ export class PlayerCore {
     #options: PlayerOptions
     #el: HTMLMediaElement
     #processor: Processor = NullProcessor;
-    #eventEmitter: PlayerEventEmitter = new PlayerEventEmitter()
+    #playerEvent: PlayerEvent = new PlayerEvent()
     get el() { return this.#el }
     constructor(el: HTMLMediaElement | string, options?: Partial<PlayerOptions>) {
         this.#options = new PlayerOptions(options ?? { minBufferTime: 13 })
         this.#el = typeof el === "string" ? document.getElementById(el) as HTMLMediaElement : el;
         if (this.#el instanceof HTMLMediaElement) {
             this.#el.addEventListener("loadedmetadata", () => {
-                this.#processor?.sourceBufferUpdate(this.#el!.currentTime);
+                //console.log("loadedmetadata");
+                
+                //this.#processor?.sourceBufferUpdate(188);
+            })
+            this.#el.addEventListener("progress", () => {
+               // this.#processor?.sourceBufferUpdate(this.#el!.currentTime);
             })
             this.#el?.addEventListener("timeupdate", throttle(() => {
-                this.#processor?.sourceBufferUpdate(this.#el!.currentTime);
-            }, 2000));
+              //  this.#processor?.sourceBufferUpdate(this.#el!.currentTime);
+            }, this.#options.sourceBufferUpdateMinFrequency * 1000));
             this.#el.addEventListener("seeking", () => {
+                console.log("seeking--11");
+                
                 this.#processor?.sourceBufferUpdate(this.#el!.currentTime);
             });
         }
     }
     /** 添加事件 */
     on<T extends ValidEvents>(event: T, listener: Listener<T>) {
-        this.#eventEmitter.on(event, listener)
+        this.#playerEvent.on(event, listener)
     }
     /** 移除事件 */
     off<T extends ValidEvents>(event: T, listener: Listener<T>) {
-        this.#eventEmitter.off(event, listener)
+        this.#playerEvent.off(event, listener)
     }
     /** 添加一次性事件 */
     once<T extends ValidEvents>(event: T, listener: Listener<T>) {
-        this.#eventEmitter.once(event, listener)
+        this.#playerEvent.once(event, listener)
     }
 
     /** 装载 MPD文件 | 分段MP4文件 异步*/
@@ -71,12 +79,12 @@ export class PlayerCore {
             const playOptions = new PlayerOptions(Object.assign(Object.assign({}, this.#options), options))
             const response = await Promise.resolve(
                 (result instanceof URL || typeof result === "string") && URL.canParse(result) ? fetch(result, { method: "HEAD" }) : result
-            )
+            )            
             if (response instanceof Response && response.ok || typeof response === "object") {
-                for await (const processor of processorList.values().map(p => p(response, this.#el, playOptions, this.#eventEmitter))) {
+                for await (const processor of processorList.values().map(p => p(response, this.#el, playOptions, this.#playerEvent.emit))) {
                     if (isProcessor(processor)) {
                         this.#processor = processor                         
-                        return (repType: RepType) => getRepList(repType, this.#processor)
+                        return (repType: RepType) => getRepList(repType, this.#processor,this.#el!.currentTime)
                     }
                 }
             }
@@ -85,7 +93,7 @@ export class PlayerCore {
         throw new PlayerError(1, "el 为空或者不为HTMLMediaElement类型", this.#el)
     }
 
-    get(repType: RepType) { return getRepList(repType, this.#processor) }
+    get(repType: RepType) { return getRepList(repType, this.#processor,this.#el!.currentTime) }
 
     /** 处理器 遍历器 */
     static processorEntries() { return processorList.entries() }