+++ lib/Player-Tools.ts
@@ -1,9 +1,9 @@
 /** 防抖 */
-export const debounce = <Fn extends (...args: any) => void>(callback: Fn, delay: number = 200) => {
+export const debounce = <Fn extends (...args: any) => any>(callback: Fn, delay: number = 200) => {
     let t: number | any
     return (...e: Parameters<Fn>) => {
         clearTimeout(t)
-        t = setTimeout(() => { callback(e) }, delay);
+        t = setTimeout(() => { callback(...e) }, delay);
     }
 }
 /** 节流 */
@@ -18,11 +18,16 @@ export const throttle = <Fn extends (...args: any) => void>(callback: Fn, durati
     }
 }
 
-/** 解析为正整数 */
-export const parsePositiveInteger = (s: string, illegal = 0) => {
-    illegal = typeof illegal === "number" ? illegal : 0
+/**
+ * 解析为正整数 
+ * @param s 要解析的字符串
+ * @param defaultValue 解析失败时的默认值
+ * @returns 
+ */
+export const parsePositiveInteger = (s: string, defaultValue = 0) => {
+    defaultValue = typeof defaultValue === "number" ? defaultValue : 0
     const num = Number.parseInt(s)
-    return Number.isInteger(num) && num >= 0 ? num : illegal
+    return Number.isInteger(num) && num >= 0 ? num : defaultValue
 }
 /** 解析一个字符串到 Document*/
 export const parseFromString = (() => {