+++ lib/Processor/MPDMSE/Segment.ts
@@ -1,14 +1,15 @@
 import { Representation } from "@lib"
 
-type RangeTimeURLType = Readonly<{
+/** 范围时间段文件类型 */
+export type RangeTimeFileType = Readonly<{
     startTime: number,
     endTime: number,
-    url: string
+    file: string
 }>
-const parseInitialization = (repid: string, segmentElement?: Element | null) => {
+const parseInitialization = (repid: string, segmentElement?: Element | null) : RangeTimeFileType | undefined => {   
     const attributeValue = segmentElement?.getAttribute("initialization") ||
         segmentElement?.getElementsByTagName("Initialization")?.[0].getAttribute("sourceURL")
-    return attributeValue ? attributeValue.replace("$RepresentationID$", repid) : undefined
+    return attributeValue ? {startTime:0,endTime:0,file: attributeValue.replace("$RepresentationID$", repid)} : undefined
 }
 
 const parseSegmentFromRangeTimeStringURLs = (() => {
@@ -25,7 +26,7 @@ const parseSegmentFromRangeTimeStringURLs = (() => {
         const timescale = parseInt(segmentTemplateElement.getAttribute("timescale") ?? "")
         let startNumber = parseInt(segmentTemplateElement.getAttribute("startNumber") ?? "")
         const sElement = Array.from(segmentTemplateElement.getElementsByTagName("S"))
-        const rangeTimeURLs = new Array<RangeTimeURLType>();
+        const rangeTimeURLs = new Array<RangeTimeFileType>();
         if (Number.isFinite(timescale) && Number.isInteger(startNumber) && mediaPresentationDuration && mediaRegExecArray && media) {
             const mediaRegExecNumber = parseInt(mediaRegExecArray[1])
             let duration = divideAndRound(segmentTemplateElement.getAttribute("duration"), timescale)
@@ -34,7 +35,7 @@ const parseSegmentFromRangeTimeStringURLs = (() => {
                     rangeTimeURLs.push(Object.freeze({
                         startTime: startNumber * duration - duration,
                         endTime: startNumber * duration,
-                        url: media.replace(mediaRegExecArray[0], (startNumber.toString().padStart(mediaRegExecNumber, "0")))
+                        file: media.replace(mediaRegExecArray[0], (startNumber.toString().padStart(mediaRegExecNumber, "0")))
                     }))
                 }
             } else if (sElement.length > 0) {
@@ -48,49 +49,50 @@ const parseSegmentFromRangeTimeStringURLs = (() => {
                             rangeTimeURLs.push(Object.freeze({
                                 startTime: duration,
                                 endTime: sduration + duration,
-                                url: media.replace(mediaRegExecArray[0], (startNumber++).toString().padStart(mediaRegExecNumber, "0"))
+                                file: media.replace(mediaRegExecArray[0], (startNumber++).toString().padStart(mediaRegExecNumber, "0"))
                             }))
                             duration += sduration;
                         }
-                    } else { return new Array<RangeTimeURLType>() }
+                    } else { return new Array<RangeTimeFileType>() }
                 }
             }
         }
         return rangeTimeURLs
     }
     const parseSegmentListFromRangeTimeStringURLs = (repid: string, segmentTemplate: Element) => {
-        return new Array<RangeTimeURLType>()
+        return new Array<RangeTimeFileType>()
     }
-    return (repid: string, mediaPresentationDuration: number, segmentElement?: Element | null): Array<RangeTimeURLType> => {
+    return (repid: string, mediaPresentationDuration: number, segmentElement?: Element | null): Array<RangeTimeFileType> => {
         if (segmentElement?.localName === "SegmentTemplate") {
             return parseSegmentTemplateFromRangeTimeStringURLs(repid, mediaPresentationDuration, segmentElement)
         } else if (segmentElement?.localName === "SegmentList") {
             return parseSegmentListFromRangeTimeStringURLs(repid, segmentElement)
         }
-        return new Array<RangeTimeURLType>()
+        return new Array<RangeTimeFileType>()
     }
 })()
 
 /** 分段文件类 */
 export class Segment {
-    #rangeTimeURLs: Array<RangeTimeURLType>
-    #initialization?: string
+    #rangeTimeURLs: Array<RangeTimeFileType>
+    #initialization?: RangeTimeFileType
     get initialization() { return this.#initialization }
     constructor(representation: Representation, segmentElement?: Element | null,) {
-        this.#initialization = parseInitialization(representation.id, segmentElement);
+        this.#initialization = parseInitialization(representation.id, segmentElement);        
         this.#rangeTimeURLs = parseSegmentFromRangeTimeStringURLs(representation.id, representation.duration, segmentElement)
     }
     /** 时间是否为属于最后一个分段文件范围内 */
-    isLastFile(time: number) {
-        const r = this.#rangeTimeURLs.at(-1);
-        return r ? (time >= r.startTime && time <= r.endTime) ? true : false : true
+    isLastFile(rangeTimeOrTime: number | RangeTimeFileType) {
+        const r = this.#rangeTimeURLs.at(-1);        
+        return r ? (typeof rangeTimeOrTime === "number"  && (rangeTimeOrTime >= r.endTime || (rangeTimeOrTime >= r.startTime && rangeTimeOrTime <= r.endTime))) ||  rangeTimeOrTime === r ? true : false: true
     }
-    getSegmentFiles(startTime: number, endTime: number) {
-        const urls = new Array<string>()
+    /** 获取时间段范围文件 */
+    getRangeTimeFiles(startTime: number, endTime: number) {
+        const urls = new Array<RangeTimeFileType>()
         for (const r of this.#rangeTimeURLs) {
             if (r.startTime > endTime) break
             if (startTime <= r.endTime && endTime >= r.startTime) {
-                urls.push(r.url)
+                urls.push(r)
             }
         }
         return urls