+++ lib/Processor/MPDMSE/index.ts
@@ -1,6 +1,6 @@
 import { PlayerOptions, PlayerEventEmitter, SwitchRepOptions } from "@lib"
 import type { Representation, Processor, ProcessorType, RepType } from "@lib"
-import { Converter, MPDConverter } from "./Converter"
+import { MPDDefaultConverter, MPDConverter } from "./Converter"
 import { SourceBufferTaskCollection } from "./SourceBufferTaskCollection"
 export { type MPDConverter }
 /**
@@ -8,31 +8,32 @@ export { type MPDConverter }
  */
 class MPDMSE implements Processor {
     #SourceBufferTaskCollection: SourceBufferTaskCollection
-    get(repType: RepType) { 
+    get(repType: RepType) {
         return this.#SourceBufferTaskCollection.getORset(repType).toArray()
     }
-    switch(repType: RepType, rep: Representation, options: SwitchRepOptions) {
-
-        this.#SourceBufferTaskCollection.changeType(repType,rep,options)
+    switch(repType: RepType, rep: Representation, currentTime:number, options: SwitchRepOptions) { 
+        this.#SourceBufferTaskCollection.changeType(repType, rep,currentTime, options)
     }
     sourceBufferUpdate(currentTime: number) { this.#SourceBufferTaskCollection.sourceBufferUpdate(currentTime) }
     constructor(sourceBufferTaskCollection: SourceBufferTaskCollection) {
         this.#SourceBufferTaskCollection = sourceBufferTaskCollection
     }
- 
+
 }
 
 const processor: ProcessorType = {
     name: "MPDMSE",
     asyncFunctionProcessorInstance: async (result: unknown, el: HTMLMediaElement, options: PlayerOptions, eventEmitter: PlayerEventEmitter) => {
-        const mpdConverter = Converter.parse(result instanceof Response ? result.url : result)
-        if(mpdConverter instanceof Converter ){
-            return  mpdConverter.asyncResponse().then(response => {
-                return response.ok  ? Promise.resolve((new SourceBufferTaskCollection(mpdConverter, el, options, eventEmitter)).sourceopen()).then(s=>{
-                    return  new MPDMSE(s)
-                }) :undefined 
-            }) 
-        }      
+        const mpdConverter = MPDDefaultConverter.parse(result instanceof Response ? result.url : result)
+        if (mpdConverter instanceof MPDDefaultConverter) {
+            return mpdConverter.asyncResponse().then(response => { 
+                return response.ok ? (new SourceBufferTaskCollection(mpdConverter, el, options, eventEmitter)).asyncSourceopen(response).then(s => {
+                   // console.log("mpdmse",s,response);
+                   // console.log(response);
+                    return new MPDMSE(s)
+                }).catch(()=>undefined) : undefined
+            })
+        }
         return undefined
     }
 }