+++ src/main.ts
@@ -1,6 +1,5 @@
 import './style.css'
-import { MPDConverter, Player, PlayerOptions } from "../lib/Player";  
-const convertedResultMap = new Map<string, URL>()
+import { MPDConverter, Player } from "../lib/Player";   
 const pl = new Player("video2") 
 const url = new URL(window.location.href + "/dash2/output.mpd?fwe")
 const urlavs = new URL(window.location.href + "/dash2/output.mpd?fwe")
@@ -10,54 +9,20 @@ pl.on("loadedmetadata",(e)=>{
 console.log("loadedmetadata--event",e);
 
 })
-pl.loaderAsync<MPDConverter>({mpd:url},{minBufferTime:120}).then(c=>{
+pl.loaderAsync<MPDConverter>({mpd:url},{minBufferTime:20}).then(c=>{
     console.log("200 ok", c("video"));
          c("video")?.at(0)?.switch()
-         //c("audio")?.at(0)?.setRep()
-         const au= c("audio")?.at(0)
-         au?.switch()
-        
+         c("audio")?.at(0)?.switch()        
     if(bs){
         c("video")?.forEach(v=>{
             const bu = document.createElement("button")
-            bu.onclick = ()=>{
-                v.switch({switchMode:"soft"})
-                au?.switch()            }
+            bu.onclick = ()=>{v.switch({switchMode:"soft"})}
             bu.innerText = `id:${v.id} b:${v.bandwidth} `
-            bs.appendChild(bu)
-          //  bs.innerHTML +=  `<div><button>${v.id }</button></div>`
-    
+            bs.appendChild(bu) 
         })
     }
-})
-const avsConverter = async(keys:string[])=>{ 
-            const urlMap = new Map<string,URL>()   
-            if(Array.isArray(keys)){
-                const fetchURL = new URL(urlavs)             
-                fetchURL.searchParams.set("returnMode","entries")
-                console.log("fetcurl",fetchURL);  
-                keys.forEach(k=>{ 
-                    if(convertedResultMap.has(k)){
-                        urlMap.set(k,convertedResultMap.get(k)!)
-                    }else{
-                        fetchURL.searchParams.append("file",  k)
-                    }                       
-                })
-                if(fetchURL.searchParams.has("file")){                  
-                    await  fetch(fetchURL).then(async r=>{
-                        if(r.ok){                            
-                            const reMap= new Map<string,string>(await r.json()); 
-                            reMap.forEach((v,keys)=>{
-                                urlMap.set(keys,new URL(v));
-                                convertedResultMap.set(keys,new URL(v));
-                             }) 
-                    }})
-                }
-            }
-            return  urlMap
-} 
+})  
 
- 
 
 // setTimeout(() => {
 //     console.group("--------------------------------");
@@ -82,9 +47,9 @@ const avsConverter = async(keys:string[])=>{
 // pl.el?.addEventListener("durationchange", () => { console.log("durationchange  当音频/视频的时长已更改时触发。",pl.el?.duration) })
 // pl.el?.addEventListener("emptied", () => { console.log("emptied  当目前的播放列表为空时触发。") })
 // pl.el?.addEventListener("ended", (e) => { console.log("ended  当目前的播放列表已结束时触发。") })
-//pl.el?.addEventListener("error", (e) => { console.log("error  当在音频/视频加载期间发生错误时触发。", e, (e.target as HTMLVideoElement).error?.message)})
+// pl.el?.addEventListener("error", (e) => { console.log("error  当在音频/视频加载期间发生错误时触发。", e, (e.target as HTMLVideoElement).error?.message)})
 // pl.el?.addEventListener("loadeddata", () => { console.log("%c loadeddata  当浏览器已加载音频/视频的当前帧时触发。") })
-// pl.el?.addEventListener("loadedmetadata", () => {console.log("%cloadedmetadata	当浏览器已加载音频/视频的元数据时触发。","color:red;");})
+//  pl.el?.addEventListener("loadedmetadata", () => {console.log("%cloadedmetadata	当浏览器已加载音频/视频的元数据时触发。","color:red;");})
 // pl.el?.addEventListener("loadstart", () => { console.log("loadstart  当浏览器开始查找音频/视频时触发。") })
 // pl.el?.addEventListener("play", () => { console.log("play  当音频/视频已开始或不再暂停时触发。") })
 // pl.el?.addEventListener("pause", () => { console.log("pause  当音频/视频已暂停时触发。") })