+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Atc-videoplayer dash流播放器 变更日志
 
+
+
+### 1.2.0-alpha.1 
+ 
+MPDMSE FetchSchedule 完善功能  
+增加（转换URL缓存 Map集合，上次差异化响应 Map集合，缓存转换URLsMap，转换列队获取响应 ，缓存转换URLsMap，缓存响应 Map，获取响应 Map集合）
+
+MPDMSE SourceBufferTask sourceBufferUpdate调整参数 
+
 ### 1.1.0-alpha.1 
 
 增加Player-Processor，Player-Representation，Player-Tools，Player-Event类型定义和相关方法